feat(pg-query-context): add callback-based withPgClient API
Expand pg-query-context with a withPgClient<T>(pool, context, fn, opts?)
function for multi-query RLS transactions. This is the callback-based
counterpart to the existing single-query pgQueryContext() API.

express-context now re-exports withPgClient from pg-query-context
instead of maintaining its own duplicate implementation.

Also exports ExecOptions and WithPgClientOptions interfaces.
