+++ postgres/pg-query-context/src/index.ts
@@ -1,5 +1,7 @@
 import { ClientBase, Pool, PoolClient, QueryResult } from 'pg';
 
+// --- Internal helpers ---
+
 function setContext(ctx: Record<string, string>): { query: string; values: string[] }[] {
   return Object.keys(ctx || {}).reduce<{ query: string; values: string[] }[]>((m, el) => {
     m.push({ query: 'SELECT set_config($1, $2, true)', values: [el, ctx[el]] });
@@ -14,7 +16,9 @@ async function execContext(client: ClientBase, ctx: Record<string, string>): Pro
   }
 }
 
-interface ExecOptions {
+// --- Single-query API (original) ---
+
+export interface ExecOptions {
   client: Pool | ClientBase;
   context?: Record<string, string>;
   query: string;
@@ -22,7 +26,7 @@ interface ExecOptions {
   skipTransaction?: boolean;
 }
 
-export default async ({ client, context = {}, query = '', variables = [], skipTransaction = false }: ExecOptions): Promise<QueryResult> => {
+async function pgQueryContext({ client, context = {}, query = '', variables = [], skipTransaction = false }: ExecOptions): Promise<QueryResult> {
   const isPool = 'connect' in client;
   const shouldRelease = isPool;
   let pgClient: ClientBase | PoolClient | null = null;
@@ -50,4 +54,49 @@ export default async ({ client, context = {}, query = '', variables = [], skipTr
       pgClient.release();
     }
   }
-};
+}
+
+export default pgQueryContext;
+
+// --- Callback-based API ---
+
+export interface WithPgClientOptions {
+  skipTransaction?: boolean;
+}
+
+/**
+ * Execute a callback within a tenant-scoped RLS transaction.
+ *
+ * Acquires a client from the pool, applies pgSettings via set_config
+ * (scoped to the transaction), calls the callback, then commits or
+ * rolls back. The client is always released back to the pool.
+ *
+ * Use this when you need to run multiple queries within the same
+ * RLS context (e.g., auth check + data mutation).
+ */
+export async function withPgClient<T>(
+  pool: Pool,
+  context: Record<string, string>,
+  fn: (client: PoolClient) => Promise<T>,
+  opts: WithPgClientOptions = {},
+): Promise<T> {
+  const client = await pool.connect();
+  try {
+    if (!opts.skipTransaction) {
+      await client.query('BEGIN');
+    }
+    await execContext(client, context);
+    const result = await fn(client);
+    if (!opts.skipTransaction) {
+      await client.query('COMMIT');
+    }
+    return result;
+  } catch (err) {
+    if (!opts.skipTransaction) {
+      await client.query('ROLLBACK').catch(() => {});
+    }
+    throw err;
+  } finally {
+    client.release();
+  }
+}
